Move logic into late-loading class
Should fix intave#1 as intave api classes are not referenced from main class anymore
